requirements.txt
torch>=2.0.0
torchvision>=0.15.0
pytorch-lightning>=2.0.0
hydra-core>=1.3.0
numpy>=1.21.0
pillow>=8.0.0
matplotlib>=3.4.0
pandas>=1.3.0
scikit-image>=0.18.0
tqdm>=4.62.0
gdown>=4.6.0 